refactor(cache): move ByteBoundedLruCache to util (pure move, no behavior change)
Extracts the count-target + byte-backstop LRU wrapper built by #4804 out of
`contract::executor` into `util::byte_bounded_lru` so the interest manager's
delta cache (#4805) can reuse the same primitive rather than growing a second
copy of the same byte accounting.

Pure move: the executor's diff is the moved code removed plus exactly one
import line. The struct, its methods, CACHE_ENTRY_OVERHEAD_BYTES and the
seven unit tests are byte-identical apart from visibility (`pub(crate)`) and
one doc sentence that pointed at "the module-level comment above", now a
path link. No call site changes.
