+++ crates/core/src/util.rs
@@ -1,4 +1,5 @@
 pub(crate) mod backoff;
+pub(crate) mod byte_bounded_lru;
 pub mod deterministic_select;
 pub(crate) mod rate_limit_layer;
 /// Order-independent `tracing` capture for tests that assert on emitted events.
